Protect the `@` when needed
In order to input gregoriotex-gsp-default.tex the `@` cannot be a protected character (it must be a normal letter).  This is automatically the case if we are in the preamble (i.e. when the defaults are initially loaded), but if we're in the document body (i.e. the user wants to revert to the defaults after having changed some distances), then we need to apply `\makeatletter` (and revert with `\makeatother`) ourselves.  The logic added here detects whether we're in the preamble and only calls `\makeatletter` and `\makeatother` if we are not.