Refactor initial height calculation
This makes the height placement of the initial a function of the number of lines which wrap it rather than a fixed amount.  It doesn't change anything for the 1 or 2 line initials, just makes it so that when 3 (or more) line initials are implemented, this calculation will be ready for them.