Refactoring to make more use of staffheight where appropriate
In line with 810631625b17f5f5b04e18e1f0e818d170b0f471 these are calculations which should have been using staffheight but weren't.
Also found another off-by-one error in the placement of the above lines text.  This was hidden by the fact that the size of the lyrics wasn't being taken into account.
The change in gregoriotex-signs.tex did not have an off-by-one error because we're drawing the ledger line exactly one interstafflinespace above the staff.  Hence, when we change to using staffheight, we still need one more factor of interstafflinespace