Modify vowel message
There is no Latin vowel definitions in gregorio-vowels.dat by default because they are built into the binary itself.  The user can, however, define their own alternative, so it is conceivable that there might be.  We thus print a little extra information when were falling back to the internal Latin vowel definitions so that the user isn't confused by the "Could not find" message.