Name font calibration distances
There were several "magic distances" still in the code.  All of these distances are related to calibrating things to the fonts and so should not be user editable.  However, having those numbers scattered through the codebase is harder to maintain (if you change one, you have to figure out which of the others also needs to change, by how much, and where they all are), so based on relationships between the numbers, I've reduced them to 6 named dimensions which are then used as needed in the various calculations.