Remove `noteadditionalspacelinestext`
The additional distance added between the score and the lyrics for lower than "normal" notes (as defined by `noteadditionalspacelinestextthreshold`) is now calculated automatically based on the exact distance below the "normal" position that the low notes have been placed.  This mirrors what was already being done for notes higher than the "normal" position above the staff.
The distance is marked as obsolete (skipping deprecated status) because it is no longer functional and we normally require deprecated code to be functional.