Refactor installation of extra doc files to TDS
We rename the variable holding the file names to make it clearer what these files are (additional doc files that need to be put into the doc folder of the TDS).
We also use the install_to function instead of a cp command (to be more consistent with the rest of the script).