Hand more of initial placement calculation over to Lua
For each line, baselineskip forms a minimum line height, so we need to check against that on a line by line basis and adjust our initial placement by that if it's bigger than the nominal height of the line.  Since we're already looping over the lines in Lua, this is much easier on the Lua side, but that means Lua needs to look at the fixed height stuff too, so while we still calculate the base amount in TeX, we no longer add it directly to the initial position adjustment amount there.