Remove alias definition of character
Most `\GreCP` macros are defined during the loading of the fonts (and thus don't show up in the TeX files).  This one is different because not all fonts have this character, so we define an alias for it which takes an argument that tells us what the alternative glyph is when its not available.  As with all `\GreCP` macros, the c files assemble the macro name piecemeal.