Update CTAN build target
The team at CTAN requested some modifications to our upload.  In particular there was an inconsistency between the zip file in package tree and in the tds zip file (some files were excluded from one but not the other, and still other files were doubled in one).  This revised build process ensures that the two files are identical.

Further, since we submit zips rather than tarballs to CTAN, I've revised the make dependencies to depend on dist-zip so that we don't have to untar and then zip the same folder as part of the process.