Normalizing filenames for Windows
Akira Kakuto reported a bug where filenames were using the wrong file separator on Windows in the TL25 Pretest, resulting in gtex files not being input properly (as part of the filename looked like a macro to TeX).  By normalizing the names just before the input step, we make sure the filename is in the format that TeX expects.

This is a bug in how the output directory was implemented.  Since that item is still in the CHANGELOG for this version, no new CHANGELOG entry is needed.