Remove Windows installer build files from distribution
Files in the windows directory are for the purpose of building the Windows installer and should not be part of the regular distribution.  For some reason, however, gregorio.ico and gregorio-resources.rc were ending up in the dist file (inside a windows directory).  This dist-hook target is automatically executed after the distribution is assembled but before it is compressed and will remove the windows directory entirely.